week-1-ts-ecosystem/w01e03-object-oriented-structure/src/Cart.ts
import Product, {ProductId, ProductType, UpdateProductParams} from "./Product";

type Products<TType extends ProductType> = Map<ProductId, Product<TType>>;

class Cart<TType extends ProductType> {
    #products: Products<TType> = new Map();

    addProduct (product: Product<TType>): void {
        this.products.set(product.id, product);
    }

    getProduct (productId: ProductId): Product<TType> | undefined {
        return this.products.get(productId);
    }

    updateProduct (id: ProductId, params: UpdateProductParams<TType>): void {
        const product = this.products.get(id);
        product.update(params);
        this.products.set(id, product);
    }

    deleteProduct (productId: ProductId): void {
        this.products.delete(productId);
    }

    get products (): Products<TType> {
        return this.#products
    }

    countProducts (): number {
        return this.products.size;
    }

    sumProducts (): number {
        const arr: [string, Product<TType>][] = Array.from(this.products);

        return arr.reduce((sum, [, product]) => {
            if (!product.price) {
                return sum;
            }

            sum += product.price * product.count;
            return sum;
        }, 0);
    }
}

export default Cart;